feat: Support to install by a specific version (#444)
This supports to install the browser by a specific version.  It is helpful to check the compatibility between major versions.  User can specific the version by a major version like `119`, or more specific version including minor, patch, and build version like `120.0.6099` or `121.0.6100.0`.  Available versions are published in [Chrome for Testing][] and the version is resolved from them.

[Chrome for Testing]: https://googlechromelabs.github.io/chrome-for-testing/