feat: Add `install-dependencies` parameter (#521)
A new parameter `install-dependencies` is added to the action. The
action will install the dependencies if the parameter is set to `true`
on the Linux. It would be useful when you use the action on the
self-hosted runner and want to install the dependencies automatically.
This feature works only on the Linux runner, otherwise, it will be
ignored.