fix: tool-cache does not work (#533)
The tool-cache package does not work for the Chromium version. It works
only for the server format. The chrome has a four-part version like
`120.0.6099.109`, and the tool-cache does not retrieve the cached
version. This pull request replaces the standard tool-cache package with
the cache that supports the chrome version format, four-part, release
channel name, snapshot number, and `latest`.

Close #504