feat!: Use chrome for testing as default (#632)
This PR introduces significant changes to improve stability and
maintainability by switching to Chrome for Testing as the default
browser.

### 1. Installer Changes on Channel Name Installs

This update replaces previous Chromium builds with the official Chrome
for Testing browser on channel name installs. The installation process
is simplified by using a standardized Chrome for Testing distribution.
The action previously installed the browser to `C:\Program Files` using
the official installer on Windows, or to `$RUNNER_TOOL_CACHE` on Linux
and macOS with the official archive. Now, the action will install the
browser to `$RUNNER_TOOL_CACHE` from [Chrome for Testing][].

[Chrome for Testing]:
https://googlechromelabs.github.io/chrome-for-testing/

### 2. Cache Directory Changes

This update introduces a new cache directory structure for the Chrome
for Testing browser. The cache directory is now named `chrome` instead
of `chromium`.

### 3. Default Version Changes

Previously, the action installed the *latest* snapshot version. Now, the
action will install the *stable* version of Chrome for Testing by
default.

## Migration Guide

### 1. Installer Changes on Channel Name Installs

If you are using the channel name (`stable`, `beta`, `dev`, or `canary`)
to install the browser, the new version of the action installs the
browser from [Chrome for Testing][] instead of the official installer.
Please check if your workflow is using the expected browser binary
version with the new version of the action.

The installed browser binary with the channel name is as follows:

| Platform | New Binary Path |
|----------|----------------|
| Linux    | `chrome` |
| macOS    | `Google Chrome for Testing` |
| Windows  | `chrome.exe` |

On Windows runner, the installed location is changed from `C:\Program
Files` to `$RUNNER_TOOL_CACHE`.

### 2. Cache Directory Changes

If you are self-hosted runner and installing the browser with the
channel name to speed up the runner setup, please update your cache
directory on your runners. The cache directory has been changed from
`chromium` to `chrome` and ensure that your workflow is using the new
cache directory.

### 3. Default Version Changes

If you don't specify `chrome-version` parameter in your workflow, please
ensure that your workflow will install the expected version of Chrome
for Testing. The default version is the *stable* channel version of
Chrome for Testing instead of the *latest* version. To install the
*latest* version, you need to specify `chrome-version: latest` in your
workflow.

```yaml
# v1
- name: Install the stable version of Chrome for Testing
  uses: browser-actions/setup-chrome@v1
  with:
    chrome-version: stable
- name: Install the latest version of Chrome for Testing
  uses: browser-actions/setup-chrome@v1

# v2
- name: Install the stable version of Chrome for Testing
  uses: browser-actions/setup-chrome@v2

- name: Install the latest version of Chrome for Testing
  uses: browser-actions/setup-chrome@v2
  with:
    chrome-version: latest
```